requirements.txt
Pillow
amazoncaptcha
beautifulsoup4
black
langchain
langchain-pinecone
langchain_community
langchain_openai
langchainhub
langgraph
lxml
nest_asyncio
pandas
playwright
playwright-stealth
python-dotenv
requests
streamlit
setuptools